Reworked USB, added power LED & cap
The USB conection used to have dangling stubs, due to the likely
unpopulated direct connection to the CC2538 native USB. It has now been
reworked to use 0402 components, and the series resistor now rests
directly on the vias, minimizing the amount of stub. Although it's just
low speed USB, I doubt it would affect much.

Additionally, wired up the activity LED (PB4) to the system's WLAN
indicator, and added a power LED (PB1), which would be turned on by the
module to indicate it's alive.

Also added a capacitor near the module's 3V3 connection.