Reduce thermal spokes to 0.3mm
Found it difficult to solder the GND pins of the Zigbee module. I'm not
sure if the heat was being conducted away by the GND plane on the
castellated GND pads of the module, or by the GND plane on this board.

In either case, it's safer to just reduce the spoke widths of the copper
pours, since 0.508mm for thermal spokes is an overkill anyway. In
addition, thermal clearance is also increased to pull the pours away
from the pad, hopefully reducing thermal conduction.